refactor: validate 함수와 inputUtil 함수들을 분리하여  모듈화
